requirements.txt
websockets>11.0.2,<12.0
requests==2.31.0
alibabacloud_dingtalk
dingtalk_stream
loguru
python-dotenv==1.0.1
PyYAML==6.0.1
sseclient-py==1.8.0
